doc reduction funcs
